+++ ckanext/harvest/logic/auth/update.py
@@ -93,6 +93,14 @@ def harvest_job_abort(context, data_dict):
     return harvest_jobs_run(context, data_dict)
 
 
+def harvest_job_terminate(context, data_dict):
+    '''
+        Authorization check for terminating a running harvest job
+        Same permissions as running one
+    '''
+    return harvest_jobs_run(context, data_dict)
+
+
 def harvest_sources_reindex(context, data_dict):
     '''
         Authorization check for reindexing all harvest sources